+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 Mapbox welcomes participation and contributions from everyone.
 
+# main
+## Bug fixes 🐞
+* Fix wrong `onLongTouch` event detected on any map gesture after clicks on ViewAnnotation. ([1954](https://github.com/mapbox/mapbox-maps-android/pull/1954))
+
 
 # 10.11.0-rc.1 January 26, 2023
 ## Features ✨ and improvements 🏁